refactor: 로직 분리 및 패키지 변경
- 용도에 맞춰서 패키지 분리
- CarRacing에는 비즈니스 로직만 남기고 게임 실행 관련은 CarRacingExecute로 이동
