+++ src/main/kotlin/racing/dto/CarRacingInputData.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.dto
 
 data class CarRacingInputData(
     val carNames: List<String>,
