+++ src/main/kotlin/racing/dto/Winners.kt
@@ -1,4 +1,4 @@
-package racing
+package racing.dto
 
 data class Winners(
     val winnerNames: List<String>
