+++ src/test/kotlin/racing/CarRacingTest.kt
@@ -2,6 +2,8 @@ package racing
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racing.domain.CarRacing
+import racing.dto.Car
 
 class CarRacingTest {
     @Test
